move utils object to context file
