+++ substantial/workflows/context.py
@@ -1,9 +1,11 @@
 import json
+import random
+import uuid
+import logging
 from typing import TYPE_CHECKING, Any, Callable, List, Optional
 from datetime import datetime, timedelta, timezone
 
 from substantial.protos import events, metadata
-from substantial.workflows.utils import Utils
 
 if TYPE_CHECKING:
     from substantial.workflows.run import Run
@@ -146,3 +148,21 @@ class Context:
         Interrupt after the call
         """
         raise CancelWorkflow(self.run)
+
+
+class Utils:
+    def __init__(self, ctx: Context):
+        self.ctx = ctx
+
+    def now(self) -> str:
+        return self.ctx.save(lambda: datetime.now(tz=timezone.utc).isoformat())
+
+    def random(self, a: int, b: int) -> int:
+        return self.ctx.save(lambda: random.randint(a, b))
+
+    def uuid4(self) -> uuid.UUID:
+        return self.ctx.save(lambda: uuid.uuid4())
+
+    @staticmethod
+    def log(level, msg, *args, **kwargs) -> None:
+        logging.log(level, msg, *args, **kwargs)
